requirements-dev.txt
pytest
pytest-mock
python-dotenv
requests
black
flake8
pandas
openpyxl
openai
python-dotenv
PyMuPDF
fitz
PyPDF2
pdf2image
pytesseract
thefuzz
google-cloud-documentai
beautifulsoup4

#DB
mysql-connector-python
SQLAlchemy

#AWS
boto3